--- src/main/java/authentication/AccessTokenProvider.java
@@ -1,53 +1,0 @@
-package authentication;
-
-import exception.UnauthorizedException;
-import exception.errorcode.ErrorCode;
-import io.jsonwebtoken.JwtException;
-import io.jsonwebtoken.Jwts;
-import io.jsonwebtoken.SignatureAlgorithm;
-import io.jsonwebtoken.security.Keys;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.stereotype.Component;
-
-import javax.crypto.SecretKey;
-import java.nio.charset.StandardCharsets;
-import java.util.Date;
-
-@Slf4j
-@Component
-public class AccessTokenProvider {
-    private final SecretKey key;
-    private final long validityInMilliseconds;
-
-    public AccessTokenProvider(@Value("${security.jwt.token.secret-access-key}") final String secretKey,
-                               @Value("${security.jwt.token.access-expire-length}") final long validityInMilliseconds) {
-        this.key = Keys.hmacShaKeyFor(secretKey.getBytes(StandardCharsets.UTF_8));
-        this.validityInMilliseconds = validityInMilliseconds;
-    }
-
-    public String createToken(final String payload) {
-        Date now = new Date();
-        Date expiration = new Date(now.getTime() + validityInMilliseconds);
-
-        return Jwts.builder()
-                .setSubject(payload)
-                .setIssuedAt(now)
-                .setExpiration(expiration)
-                .signWith(key, SignatureAlgorithm.HS256)
-                .compact();
-    }
-
-    public String getPayload(final String token) {
-        try {
-            return Jwts.parserBuilder()
-                    .setSigningKey(key)
-                    .build()
-                    .parseClaimsJws(token)
-                    .getBody()
-                    .getSubject();
-        } catch (JwtException e) {
-            throw new UnauthorizedException(ErrorCode.INVALID_TOKEN, e.getMessage());
-        }
-    }
-}
