--- src/main/java/authentication/AuthenticatedUserArgumentResolver.java
@@ -1,29 +1,0 @@
-package authentication;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.core.MethodParameter;
-import org.springframework.web.bind.support.WebDataBinderFactory;
-import org.springframework.web.context.request.NativeWebRequest;
-import org.springframework.web.method.support.HandlerMethodArgumentResolver;
-import org.springframework.web.method.support.ModelAndViewContainer;
-
-@Configuration
-@RequiredArgsConstructor
-public class AuthenticatedUserArgumentResolver implements HandlerMethodArgumentResolver {
-    private final AuthenticationContext authenticationContext;
-
-    @Override
-    public boolean supportsParameter(final MethodParameter parameter) {
-        return parameter.hasParameterAnnotation(AuthenticatedUser.class);
-    }
-
-    @Override
-    public Object resolveArgument(final MethodParameter parameter,
-                                  final ModelAndViewContainer mavContainer,
-                                  final NativeWebRequest webRequest,
-                                  final WebDataBinderFactory binderFactory) throws Exception {
-        return authenticationContext.getPrincipal();
-    }
-}
-
