+++ src/main/java/com/example/bjjm/authentication/AuthenticatedUser.java
@@ -1,0 +1,11 @@
+package com.example.bjjm.authentication;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+@Target(ElementType.PARAMETER)
+@Retention(RetentionPolicy.RUNTIME)
+public @interface AuthenticatedUser {
+}
