+++ src/main/java/com/example/bjjm/authentication/AuthenticationContext.java
@@ -1,0 +1,15 @@
+package com.example.bjjm.authentication;
+
+import com.example.bjjm.entity.User;
+import lombok.Getter;
+import lombok.Setter;
+import org.springframework.stereotype.Component;
+import org.springframework.web.context.annotation.RequestScope;
+
+@Getter
+@Setter
+@Component
+@RequestScope
+public class AuthenticationContext {
+    private User principal;
+}
