+++ src/main/java/com/example/bjjm/controller/UserController.java
@@ -1,0 +1,12 @@
+package com.example.bjjm.controller;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequiredArgsConstructor
+@RequestMapping("/users")
+public class UserController {
+
+}
