+++ src/main/java/com/example/bjjm/entity/User.java
@@ -1,0 +1,29 @@
+package com.example.bjjm.entity;
+
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.Table;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Builder
+@AllArgsConstructor
+@NoArgsConstructor
+@Entity
+@Getter
+@Table(name = "users")
+public class User extends BaseEntity {
+    @Column(nullable = false)
+    private Long kakaoId;
+
+    @Column(nullable = false)
+    private String name;
+
+//    @Column
+//    private String email;
+
+    @Column
+    private String profileImage;
+}
