+++ src/main/java/com/example/bjjm/exception/CustomException.java
@@ -1,0 +1,20 @@
+package com.example.bjjm.exception;
+
+import com.example.bjjm.exception.errorcode.ErrorCode;
+import lombok.Getter;
+
+@Getter
+public class CustomException extends RuntimeException {
+    private final ErrorCode errorCode;
+    private final String detail;
+
+    public CustomException(ErrorCode errorCode) {
+        this.errorCode = errorCode;
+        this.detail = null;
+    }
+
+    public CustomException(ErrorCode errorCode, String detail) {
+        this.errorCode = errorCode;
+        this.detail = detail;
+    }
+}
