+++ src/main/java/com/example/bjjm/repository/UserRepository.java
@@ -1,0 +1,11 @@
+package com.example.bjjm.repository;
+
+import com.example.bjjm.entity.User;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.Optional;
+import java.util.UUID;
+
+public interface UserRepository extends JpaRepository<User, UUID> {
+    Optional<User> findByKakaoId(Long kakaoId);
+}
