+++ src/main/java/com/example/bjjm/service/UserService.java
@@ -1,0 +1,10 @@
+package com.example.bjjm.service;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+
+@Service
+@RequiredArgsConstructor
+public class UserService {
+
+}
