--- src/main/java/exception/BadRequestException.java
@@ -1,10 +1,0 @@
-package exception;
-
-import exception.errorcode.ErrorCode;
-
-public class BadRequestException extends CustomException {
-    public BadRequestException(ErrorCode errorCode) { super(errorCode); }
-    public BadRequestException(ErrorCode errorCode, String detail) {
-        super(errorCode, detail);
-    }
-}
