--- src/main/java/repository/UserRepository.java
@@ -1,9 +1,0 @@
-package repository;
-
-import entity.User;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.UUID;
-
-public interface UserRepository extends JpaRepository<User, UUID> {
-}
