--- src/main/java/service/UserService.java
@@ -1,10 +1,0 @@
-package service;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-
-@Service
-@RequiredArgsConstructor
-public class UserService {
-
-}
